chore: Add more Atlassian logging
Mainly associate the logs with the traces so it's possible to check
which GraphQL request caused a certain debug output.
