+++ packages/server/graphql/types/JiraRemoteProject.ts
@@ -1,12 +1,11 @@
 import {GraphQLBoolean, GraphQLID, GraphQLNonNull, GraphQLObjectType, GraphQLString} from 'graphql'
 import JiraProjectId from 'parabol-client/shared/gqlIds/JiraProjectId'
-import {JiraProject} from 'parabol-client/utils/AtlassianManager'
 import {
   createImageUrlHash,
   createParabolImageUrl,
   downloadAndCacheImage
 } from '../../utils/atlassian/jiraImages'
-import AtlassianServerManager from '../../utils/AtlassianServerManager'
+import AtlassianServerManager, {JiraProject} from '../../utils/AtlassianServerManager'
 import {GQLContext} from '../graphql'
 import IntegrationProviderServiceEnum from './IntegrationProviderServiceEnum'
 import JiraRemoteAvatarUrls from './JiraRemoteAvatarUrls'
